feat: repository 메소드 분리
