+++ src/main/java/com/example/yeobee/core/calculation/domain/CalculationRepository.java
@@ -1,0 +1,17 @@
+package com.example.yeobee.core.calculation.domain;
+
+import com.example.yeobee.core.expense.domain.ExpenseType;
+import java.util.List;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface CalculationRepository {
+
+    List<CalculationResult> getCalculationResult(Long tripId);
+
+    Long getTotalSharedBudgetIncome();
+
+    Long getTotalSharedBudgetExpense();
+
+    List<CalculationResult> getTotalExpensePerTripUser(Long tripId, ExpenseType expenseType);
+}
