+++ src/main/java/com/example/yeobee/core/expense/domain/CustomExpenseRepositoryImpl.java
@@ -3,18 +3,14 @@ package com.example.yeobee.core.expense.domain;
 import static com.example.yeobee.core.expense.domain.QExpense.expense;
 import static com.example.yeobee.core.trip.domain.QTrip.trip;
 import static com.example.yeobee.core.trip.domain.QTripCurrency.tripCurrency;
-import static com.example.yeobee.core.trip.domain.QTripUser.tripUser;
 
 import com.example.yeobee.common.exception.BusinessException;
 import com.example.yeobee.common.exception.ErrorCode;
-import com.example.yeobee.core.calculation.domain.CalculationResult;
 import com.example.yeobee.core.expense.dto.request.ExpenseListFilter;
 import com.example.yeobee.core.expense.dto.response.ExpenseListRetrieveResponseDto;
 import com.querydsl.core.types.Predicate;
-import com.querydsl.core.types.Projections;
 import com.querydsl.core.types.dsl.BooleanExpression;
 import com.querydsl.jpa.impl.JPAQueryFactory;
-import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.LocalTime;
@@ -50,48 +46,6 @@ public class CustomExpenseRepositoryImpl implements CustomExpenseRepository {
             .fetchOne();
         if (count == null) throw new BusinessException(ErrorCode.EXPENSE_NOT_FOUND);
         return new PageImpl<>(expenseList.stream().map(ExpenseListRetrieveResponseDto::new).toList(), pageable, count);
-    }
-
-    @Override
-    public List<CalculationResult> getCalculationResult(Long tripId) {
-        return queryFactory.select(Projections.constructor(CalculationResult.class,
-                                                           expense.payer,
-                                                           expense.amount
-                                                               .multiply(tripCurrency.exchangeRate.value)
-                                                               .divide(tripCurrency.exchangeRate.standard)
-                                                               .sum().coalesce(BigDecimal.ZERO)))
-            .from(expense)
-            .leftJoin(expense.payer, tripUser)
-            .leftJoin(expense.tripCurrency, tripCurrency)
-            .where(expense.trip.id.eq(tripId)
-                       .and(expense.expenseType.eq(ExpenseType.SHARED))
-                       .and(expense.payer.isNotNull()))
-            .groupBy(expense.payer.id)
-            .fetch();
-    }
-
-    @Override
-    public Long getTotalSharedBudgetIncome() {
-        return queryFactory.select(expense.amount
-                                       .multiply(tripCurrency.exchangeRate.value)
-                                       .divide(tripCurrency.exchangeRate.standard)
-                                       .sum().coalesce(BigDecimal.ZERO).longValue())
-            .from(expense)
-            .leftJoin(expense.tripCurrency, tripCurrency)
-            .where(expense.expenseType.eq(ExpenseType.SHARED_BUDGET_INCOME))
-            .fetchOne();
-    }
-
-    @Override
-    public Long getTotalSharedBudgetExpense() {
-        return queryFactory.select(expense.amount
-                                       .multiply(tripCurrency.exchangeRate.value)
-                                       .divide(tripCurrency.exchangeRate.standard)
-                                       .sum().coalesce(BigDecimal.ZERO).longValue())
-            .from(expense)
-            .leftJoin(expense.tripCurrency, tripCurrency)
-            .where(expense.expenseType.eq(ExpenseType.SHARED).and(expense.payer.isNull()))
-            .fetchOne();
     }
 
     private Predicate[] getPredicates(ExpenseListFilter filter) {
