--- src/main/java/com/example/yeobee/core/trip/domain/CustomTripUserRepository.java
@@ -1,10 +1,0 @@
-package com.example.yeobee.core.trip.domain;
-
-import com.example.yeobee.core.calculation.domain.CalculationResult;
-import com.example.yeobee.core.expense.domain.ExpenseType;
-import java.util.List;
-
-public interface CustomTripUserRepository {
-
-    List<CalculationResult> getTotalExpensePerTripUser(Long tripId, ExpenseType expenseType);
-}
