+++ src/main/java/com/example/yeobee/core/trip/domain/TripUserRepository.java
@@ -4,7 +4,7 @@ import com.example.yeobee.core.user.domain.User;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TripUserRepository extends JpaRepository<TripUser, Long>, CustomTripUserRepository {
+public interface TripUserRepository extends JpaRepository<TripUser, Long> {
 
     Optional<TripUser> findByTripAndUser(Trip trip, User user);
 }
